requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bootstrap4==0.1.0
click==8.0.3
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-bootstrap4==3.0.1
django-environ==0.8.1
djangorestframework==3.12.4
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
rander==1.0.0
self==2020.12.3
showdata==1.4.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
temp==2020.7.2
tqdm==4.62.3
Werkzeug==2.0.2
